fix meter not supporting long-based transfer limit

+++ src/main/java/com/almostreliable/energymeter/block/component/EnergyHandlerHost.java
@@ -20,5 +20,5 @@ public interface EnergyHandlerHost {
 
     TransferMode getTransferMode();
 
-    int getTransferLimit();
+    long getTransferLimit();
 }
